:sparkles: add the new fetch policy type
we only support 'cache-first' and 'cache-and-network' at the moment. not supporting a 'network-only' option is a limitation of using selectors, because we don't know if data is cached for this particular hook, or if it's newly fetched data. We would need a way to determine what hooks are connected to a certain query.